git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4591 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWorkSim/src/rwsim/rwphysics/ContactModelFactory.cpp
@@ -68,7 +68,7 @@ namespace {
 				std::list<FilteredPoint>& blobs,
 				double sepDist)
 	{
-		for(int i=0; i<dists.p1s.size(); i++){
+		for(size_t i=0; i<dists.p1s.size(); i++){
 			const Vector3D<> &v = dists.p1s[i];
 			bool inBlob = false;
 			BOOST_FOREACH(FilteredPoint& p, blobs){
