+++ RobWorkSim/src/rwsim/sensor/TactileArraySensor.cpp
@@ -60,7 +60,7 @@ namespace {
         if(i_low<0) i_low = 0;
         if(i_upp>=(size_t)tMatrix.size1()) i_upp = tMatrix.size1()-1;
 
-        int j_low = j-1,j_upp=j+1;
+        size_t j_low = j-1,j_upp=j+1;
         if(j_low<0) j_low = 0;
         if(j_upp>=tMatrix.size2()) j_upp = tMatrix.size2()-1;
